refactor(KvWwwHeaderBasic): move composables to src/utils and rename
Relocate the header composables into the package's standard src/utils dir
and generalize the two reusable ones:
- useHeaderBasicBreakpoint -> useBreakpoints
- useHeaderBasicSearch     -> useTypeaheadSearch
- useHeaderBasicMenuState  (name kept, moved)

Move their specs to tests/unit/specs/utils/ and update all imports to
the #utils/ alias.
