+++ @kiva/kv-components/tests/unit/specs/utils/useBreakpoints.spec.ts
@@ -1,20 +1,20 @@
-import { useHeaderBasicBreakpoint } from '#components/KvWwwHeaderBasic/composables/useHeaderBasicBreakpoint';
+import { useBreakpoints } from '#utils/useBreakpoints';
 
 function setWidth(px: number) {
 	Object.defineProperty(window, 'innerWidth', { writable: true, configurable: true, value: px });
 }
 
-describe('useHeaderBasicBreakpoint', () => {
+describe('useBreakpoints', () => {
 	it('isMobile is true below the md breakpoint', () => {
 		setWidth(500);
-		const { isMobile, checkIsMobile } = useHeaderBasicBreakpoint();
+		const { isMobile, checkIsMobile } = useBreakpoints();
 		checkIsMobile();
 		expect(isMobile.value).toBe(true);
 	});
 
 	it('isMobile is false at/above the md breakpoint', () => {
 		setWidth(1200);
-		const { isMobile, checkIsMobile } = useHeaderBasicBreakpoint();
+		const { isMobile, checkIsMobile } = useBreakpoints();
 		checkIsMobile();
 		expect(isMobile.value).toBe(false);
 	});
